Corrected getCost() method.

+++ src/org/hsm/model/Plant.java
@@ -19,7 +19,7 @@ public interface Plant {
      *
      * @return the cost of the plant
      */
-    int getCost();
+    double getCost();
 
     /**
      *
